% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{luigg}
\alias{luigg}
\title{Crayons Color Palettes}
\format{
list of character vectors of type \code{c('palette', 'character')}
}
\usage{
luigg
}
\description{
Crayon colors, based on common Crayola colors.
Palettes build from \href{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}{Wikipedia}
}
\details{
Included palettes:
\itemize{
\item \strong{super_mario}, Super Mario Bros. Palette
}
}
\examples{
plot(luigg$super_mario)

}
\references{
Dionysios Politis et. al. International Journal of New Technologies in Science and Engineering
Vol. 4, Issue. 1, Jan 2017, ISSN 2349-0780
}
\concept{colors}
\keyword{datasets}
