% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{luigg}
\alias{luigg}
\title{Mario Brothers Color Palettes}
\format{
list of character vectors of type \code{c('palette', 'character')}
}
\usage{
luigg
}
\description{
Color palettes from various Mario/Nintendo games
}
\details{
Included palettes:
\itemize{
\item \strong{super_mario}, Super Mario Bros. palette (from Politis et al 2017)
\item \strong{double_dash}, Mario Kart Double Dash palette
}
}
\examples{
plot(luigg$super_mario)

}
\references{
Dionysios Politis et. al. International Journal of New Technologies in Science and Engineering
Vol. 4, Issue. 1, Jan 2017, ISSN 2349-0780
}
\concept{colors}
\keyword{datasets}
